+++ tests/test_api.py
@@ -1,12 +1,27 @@
 import os
 import subprocess
 import time
+import yaml
+import json
+import pickle
+from pathlib import Path
 
 import pytest
 
 from ipumspy import api
-from ipumspy.api import IpumsApiClient, OtherExtract, UsaExtract
-from ipumspy.api.exceptions import BadIpumsApiRequest, IpumsApiException, IpumsNotFound
+from ipumspy.api import (
+    IpumsApiClient,
+    OtherExtract,
+    UsaExtract,
+    extract_from_dict,
+    extract_to_dict,
+)
+from ipumspy.api.exceptions import (
+    BadIpumsApiRequest,
+    IpumsApiException,
+    IpumsNotFound,
+    IpumsExtractNotSubmitted,
+)
 
 
 @pytest.fixture(scope="module")
@@ -50,6 +65,7 @@ def test_usa_build_extract():
         "variables": {"AGE": {}, "SEX": {}},
         "description": "My IPUMS extract",
         "data_format": "fixed_width",
+        "collection": "usa",
     }
 
 
@@ -151,6 +167,18 @@ def test_not_found_exception(live_api_client: IpumsApiClient):
     )
 
 
+def test_not_submitted_exception():
+    extract = UsaExtract(
+        ["us2012b"],
+        ["AGE", "SEX"],
+    )
+    with pytest.raises(IpumsExtractNotSubmitted) as exc_info:
+        dct = extract_to_dict(extract)
+    assert exc_info.value.args[0] == (
+        "Extract has not been submitted and so has no json response"
+    )
+
+
 @pytest.mark.integration
 def test_extract_was_purged(live_api_client: IpumsApiClient):
     """
@@ -158,3 +186,45 @@ def test_extract_was_purged(live_api_client: IpumsApiClient):
     """
     was_purged = live_api_client.extract_was_purged(extract="1", collection="usa")
     assert was_purged == True
+
+
+def test_extract_from_dict(fixtures_path: Path):
+    with open(fixtures_path / "example_extract.yml") as infile:
+        extract = extract_from_dict(yaml.safe_load(infile))
+
+    for item in extract:
+        assert item.collection == "usa"
+        assert item.samples == ["us2012b"]
+        assert item.variables == ["AGE", "SEX", "RACE"]
+
+    with open(fixtures_path / "example_extract.json") as infile:
+        extract = extract_from_dict(json.load(infile))
+
+    for item in extract:
+        assert item.collection == "usa"
+        assert item.samples == ["us2012b"]
+        assert item.variables == ["AGE", "SEX", "RACE"]
+
+
+def test_extract_to_dict(fixtures_path: Path):
+    # reconstitute the extract object from pickle
+    with open(fixtures_path / "usa_extract_obj.pkl", "rb") as infile:
+        extract = pickle.load(infile)
+
+    # export extract to dict
+    dct = extract_to_dict(extract)
+    assert dct["collection"] == "usa"
+    assert dct["samples"] == {"us2012b": {}}
+    assert dct["variables"] == {
+        "YEAR": {"preselected": True},
+        "SAMPLE": {"preselected": True},
+        "SERIAL": {"preselected": True},
+        "CBSERIAL": {"preselected": True},
+        "GQ": {"preselected": True},
+        "HHWT": {"preselected": True},
+        "PERNUM": {"preselected": True},
+        "PERWT": {"preselected": True},
+        "AGE": {},
+        "SEX": {},
+        "RACE": {},
+    }