+++ docs/source/change-log.rst
@@ -11,8 +11,16 @@ This project adheres to `Semantic Versioning`_.
 .. _Semantic Versioning: http://semver.org/
 
 
-0.1.0
+0.*.*
 -----
 year-month-day
+
+* Added :py:meth:`~ipumspy.api.extract.extract_to_dict()` method to enable easy exporting of extract objects to dictionary objects for eventual writing to .json for ease of extract sharing.
+* Added :py:meth:`~ipumspy.api.exceptions.IpumsExtractNotSubmitted` exception
+
+
+0.1.0
+-----
+2021-11-30
 
 * This is the initial version of ipumspy.